requirements.txt
# StandX 메이커 봇 의존성
# =======================

# 네트워크
requests>=2.31.0
websockets>=12.0

# 암호화 및 인증
PyNaCl>=1.5.0           # ed25519 서명
eth-account>=0.10.0     # EVM 지갑 서명
PyJWT>=2.8.0            # JWT 처리

# 설정
python-dotenv>=1.0.0
PyYAML>=6.0.0

# 로깅
colorlog>=6.8.0

# 유틸리티
typing-extensions>=4.8.0
